==> scripts/step05_climate_data_prep.Rmd <==
---
title: "Climate Data Prep"
author: "Kristin Braziunas"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = '../')
```

## Purpose

Download climate data and prepare predictor climate variables relevant for tree regeneration.

Climate data (4 km resolution, monthly) was downloaded from Terraclimate [https://www.climatologylab.org/terraclimate.html](https://www.climatologylab.org/terraclimate.html) for plot center coordinates for 2021 field data and Schoennagel et al. 2003 field data. Downloaded variables included: climate water deficit and vapor pressure deficit.

```{r load-libs, results="hide", message=FALSE, warning=FALSE}
# load libraries
library(tidyverse)
library(openxlsx)
library(ggpubr)
```

## 1. Read in and prep climate data

Coarse resolution data was used because plot pairs were designed to represent similar conditions, and this ensured that most plot pairs occurred within the same grid cell and were therefore assigned the same average climate conditions for the purpose of analysis. Of the 33 plot pairs, 27 occurred within the same grid cell so had identical climate conditions (including all plots sampled in 2021) and 6 had differing climate conditions. Of the 6 pairs that differed, almost all (5 of the 6 pairs) differed by < 5% in the average value of each of these climate variables. 1 plot pair differed by up to 15% for some variables.

```{r prep-data, results="hide",message=FALSE,warning=FALSE}

#####
# 1. read in and prep terraclim data
#####

### read in plot coordinates

fire.final <- read.csv("processed_data/plot_selection/final_plot_coords_epsg4326.csv")

### read in files

terra.clim <- data.frame()

# loop through all plots, all variables
for(i in list.files("data/terraclim/", pattern="csv")) {
  print(i)
  file.name<-i
  # variable name
  f1 <- strsplit(file.name, split=".csv")
  var <- strsplit(f1[[1]], split="_")[[1]][2]

  # dataframe of daily values
  day.in <- read.csv(paste0("data/terraclim/",file.name), header=FALSE,skip=1)
  clim.in <- cbind(day.in,var)
  names(clim.in) <- c("time","lat","lon","value","var")
  
  # bind to main output
  terra.clim <- rbind(terra.clim,clim.in)
  
}

summary(terra.clim)

### determine scale factor, reference parent dataset
# http://thredds.northwestknowledge.net:8080/thredds/dodsC/agg_terraclimate_def_1958_CurrentYear_GLOBE.nc.html, def = 0.1 (mm)
# http://thredds.northwestknowledge.net:8080/thredds/dodsC/agg_terraclimate_vpd_1958_CurrentYear_GLOBE.nc.html, vpd = 0.01 (kPa)

### match up with plots
terra.short <- terra.clim %>%
  full_join(fire.final, by=c("lat" = "Y","lon" = "X")) %>%
  pivot_wider(names_from=var, values_from=value) %>%
  mutate(time = as.Date(time)) %>%
  mutate(year= as.numeric(format(time,'%Y')),
         month = as.numeric(format(time,'%m'))) %>%
  # apply scale factors
  mutate(vpd = 0.01*vpd, def = 0.1*def) 

head(terra.short)
summary(terra.short)

### how often do plot pairs have matching climate? (i.e., co-occur in same 4 km grid cell)
terra.match <- terra.clim %>%
  full_join(fire.final, by=c("lat" = "Y","lon" = "X")) %>%
  # mean values for each variable
  group_by(Plot_code,Fire_interval,var) %>%
  summarise(value=mean(value)) %>%
  # set up comparison of long and short pairs
  # split based on naming convention of 2021 plots
  separate(Plot_code, into=c("fire1","fire2","site","plot","fire_int_temp","id"), sep="_", fill="right") %>%
  # identify correct matches for schoennagel plots
  mutate(fire2 = ifelse(is.na(id),"fire2",fire2),
         site = ifelse(is.na(id),1,site),
         plot = ifelse(is.na(id),1,plot)) %>%
  dplyr::select(-c(fire_int_temp,id)) %>%
  pivot_wider(names_from=Fire_interval,values_from=value) %>%
  # calculate absolute and relative difference
  mutate(mean=(Short+Long)/2,
         diff=Short-Long,
         rel_diff = abs(diff/mean)*100)

# pairs with identical climates = 27 (54 plots)
terra.match %>%
  filter(diff==0) %>%
  tally()/2

# pairs with differing climates = 6 (12 plots)
terra.match %>%
  filter(diff!=0) %>%
  tally()/2

# take a closer look
terra.diff <- terra.match %>%
  filter(diff!=0) 
# all differing plots are from 2000 plots, none from 2021 field plots
# for all plots except 1, average climate differs by < 5%. 
# for 1 plot pair, average climate differs by up to ~15%

```

## 2-3. Calculate climate predictors

I then calculated the following climate drivers hypothesized to affect tree regeneration based on previous studies:

- Water-year (October-September) climate water deficit (normal 1989-2018, 3 yrs post-fire; Stevens-Rumann et al. 2018)
- Mean summer (June-Aug) vapor pressure deficit (normal 1989-2018, 3 yrs post-fire; Davis et al. 2019)

I examined annual trends to identify whether and which variables captured warming-drying trends over time.

```{r calc-clim-vars, results="hide",message=FALSE,warning=FALSE}
#####
# 2. calculate climate variables of interest
#####

# add water-year indicator, Oct-Sept
terra.short$water_year <- ifelse(terra.short$month %in% c(1:9),
                              terra.short$year,
                              terra.short$year+1)

### climate water deficit, from Harvey et al. 2016, Stevens-Rumann et al. 2018
# annual climate water deficit for complete years
terra.def_yr <- terra.short %>%
  # 1988 and 2022 are incomplete records for water year
  filter(!water_year %in% c(1988,2022)) %>%
  # annual CWD
  group_by(Plot_code,water_year) %>%
  summarise(ann_def_mm = sum(def)) %>%
  rename(year=water_year)

# 30-year normals, using 1989-2018
terra.def_norm <- terra.def_yr %>%
  filter(year %in% c(1989:2018)) %>%
  group_by(Plot_code) %>%
  # mean and standard deviation
  summarise(ann_def_mm_norm=mean(ann_def_mm),
            ann_def_norm_sd=sd(ann_def_mm))

### mean summer VPD from davis et al. 2019
terra.summer <- terra.short %>%
  # summer: june-august
  filter(month %in% c(6:8)) %>%
  dplyr::select(Plot_code,year,month,vpd) %>%
  group_by(Plot_code,year) %>%
  summarise(summer_vpd_kpa = mean(vpd)) 

# put these variables together
terra.ann <- terra.def_yr %>%
  left_join(terra.summer, by=c("Plot_code","year")) 

# 30-year normals for each of these
terra.norms <-  terra.ann %>%
  filter(year %in% c(1989:2018)) %>%
  dplyr::select(-c(year,ann_def_mm)) %>%
  group_by(Plot_code) %>%
  # mean and sd
  summarise_all(list(summer_vpd_kpa_norm=mean, summer_vpd_kpa_norm_sd=sd))

### quick look at trends
# annual trends
terra.ann %>%
  pivot_longer(cols=c(ann_def_mm:summer_vpd_kpa)) %>%
  ggplot(aes(x=year,y=value)) +
  facet_wrap(~name, scales="free_y") +
  geom_point() +
  geom_smooth(method="lm") +
  stat_cor(method="spearman",cor.coef.name = "rho", p.digits=NA, label.sep="") +
  theme_bw() # trend toward warmer, drier over time
  
  
#####
# 3. aggregate to 3 years postfire
#####

### read in fire year
# schoennagel data
scho.in <- read.csv("data/field_plots_2021/cleaned_data/Schoennagel_etal2003_cleaned.csv", header=TRUE) %>%
  # add or rename columns to align with new data
  mutate(Plot_code = paste0(SiteName,"_",SiteID)) %>%
  dplyr::select(c(Plot_code,Fire_year_2))

# new field plots
fire.plots <- read.xlsx("data/field_plots_2021/raw_data/SILI_field_data_2021.xlsx", 
                        sheet="Plot_info",
                        colNames=TRUE) %>%
  # select only needed columns
  dplyr::select(Plot_code,Fire_year_2) %>%
  rbind(scho.in) 

### loop through all fires, match up correct postfire years, calculate climate variables
terra.prep <- fire.plots %>%
  left_join(terra.ann, by=c("Plot_code"))

# output df
terra.out <- data.frame()

# loop through plots
for(j in 1:dim(fire.plots)[1]) {

  plot.n <- fire.plots[j,]$Plot_code
  fire.yr <- fire.plots[j,]$Fire_year_2

  # average of selected variables 
  # 3 yrs postfire
  mean.clim <- terra.prep %>%
    filter(Plot_code == plot.n,
           year %in% c(fire.yr+1,fire.yr+2,fire.yr+3)) %>%
    dplyr::select(-year) %>%
    group_by(Plot_code,Fire_year_2) %>%
    summarise_all(list(pf3=mean))

  terra.out <- rbind(terra.out,data.frame(mean.clim))

}

### master output for all climate variables
terra.outnorms <- terra.out %>%
  left_join(terra.def_norm, by="Plot_code") %>%
  left_join(terra.norms, by="Plot_code")

# quick look at trends
terra.outnorms %>%
  pivot_longer(cols=c(ann_def_mm_pf3,ann_def_mm_norm,summer_vpd_kpa_pf3,summer_vpd_kpa_norm)) %>%
  ggplot(aes(x=Fire_year_2,y=value)) +
  facet_wrap(~name, scales="free_y") +
  geom_point() +
  geom_smooth(method="lm") +
  stat_cor(method="spearman",cor.coef.name = "rho", p.digits=NA, label.sep="") +
  theme_bw() # post-fire deficit does not capture general warming-drying trend, but other variables do

### calculate postfire anomaly for focal synthetic vars: cwd and summer vpd
terra.def_anom <- terra.outnorms %>%
  mutate(anom_def_mm = ann_def_mm_pf3-ann_def_mm_norm,
         anom_def_zscore = anom_def_mm/ann_def_norm_sd,
         anom_summer_vpd_kpa = summer_vpd_kpa_pf3-summer_vpd_kpa_norm,
         anom_summer_vpd_zscore = anom_summer_vpd_kpa/summer_vpd_kpa_norm_sd) %>%
  # only needed columns
  dplyr::select(c(Plot_code,Fire_year_2,ann_def_mm_norm,summer_vpd_kpa_norm,anom_def_mm:anom_summer_vpd_zscore))

# quick look at trends
terra.def_anom %>%
  pivot_longer(cols=c(anom_def_zscore,anom_summer_vpd_zscore)) %>%
  ggplot(aes(x=Fire_year_2,y=value)) +
  facet_wrap(~name, scales="free_y") +
  geom_point() +
  geom_smooth(method="lm") +
  stat_cor(method="spearman",cor.coef.name = "rho", p.digits=NA, label.sep="") +
  theme_bw() # post-fire deficit does not capture general warming-drying trend, but summer vpd does

### write out - commented out for Rmd
# write.csv(terra.def_anom, "processed_data/climate/terraclim_def_vpd_anom.csv",row.names=FALSE)

```
